Removed UnnamedToken and related functionality
