+++ src/main/scala/firrtl/analyses/IRLookup.scala
@@ -265,3 +265,22 @@ class IRLookup private[analyses] ( private val declarations: mutable.LinkedHashM
     }
   }
 }
+
+
+/** Used for obtaining a tag for a given label
+  *
+  * E.g. Used for constructing unnamed tokens
+  */
+class TokenTagger {
+  private val counterMap = mutable.HashMap[String, Int]()
+  def getTag(label: String): Int = {
+    val tag = counterMap.getOrElse(label, 0)
+    counterMap(label) = tag + 1
+    tag
+  }
+
+  def getRef(label: String): String = {
+    "@" + label + "#" + getTag(label)
+  }
+}
+
